+++ R/dpmeans.R
@@ -32,14 +32,15 @@
 #' solC <- dpmeans(X, lambda=20)$cluster
 #' 
 #' ## visualize the results
-#' opar <- par(mfrow=c(1,4), pty="s", no.readonly=TRUE)
+#' opar <- par(no.readonly=TRUE)
+#' par(mfrow=c(1,4), pty="s")
 #' plot(X,col=lab,  pch=19, cex=.8, main="True", xlab="x", ylab="y")
 #' plot(X,col=solA, pch=19, cex=.8, main="dpmeans lbd=5", xlab="x", ylab="y")
 #' plot(X,col=solB, pch=19, cex=.8, main="dpmeans lbd=10", xlab="x", ylab="y")
 #' plot(X,col=solC, pch=19, cex=.8, main="dpmeans lbd=20", xlab="x", ylab="y")
 #' par(opar)
 #' 
-#' \dontrun{
+#' \donttest{
 #' ## let's find variations by permuting orders of update
 #' ## used setting : lambda=20, we will 8 runs
 #' sol8 <- list()
@@ -48,7 +49,8 @@
 #' }
 #' 
 #' ## let's visualize
-#' vpar <- par(mfrow=c(2,4), pty="s", no.readonly=TRUE)
+#' vpar <- par(no.readonly=TRUE)
+#' par(mfrow=c(2,4), pty="s")
 #' for (i in 1:8){
 #'   pm = paste("permute no.",i,sep="")
 #'   plot(X,col=sol8[[i]], pch=19, cex=.8, main=pm, xlab="x", ylab="y")